!6 пересортировал тесты по модулям.
