+++ common/src/main/java/ru/practicum/shareit/booking/dto/BookingItemDto.java
