+++ common/src/main/java/ru/practicum/shareit/item/dto/ItemWithBookingDto.java
