+++ common/src/test/java/shareit/booking/dto/BookingInDtoTest.java
@@ -1,10 +1,11 @@
-package ru.practicum.shareit.comment.dto;
+package shareit.booking.dto;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.json.JsonTest;
 import org.springframework.boot.test.json.JacksonTester;
 import ru.practicum.shareit.Constants;
+import ru.practicum.shareit.booking.dto.BookingInDto;
 
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
@@ -12,30 +13,29 @@ import java.time.format.DateTimeFormatter;
 import static org.assertj.core.api.AssertionsForInterfaceTypes.assertThat;
 
 @JsonTest
-class CommentDtoTest {
+class BookingInDtoTest {
     private final DateTimeFormatter dtf = Constants.DATE_TIME_FORMATTER;
     @Autowired
-    private JacksonTester<CommentDto> json;
+    private JacksonTester<BookingInDto> json;
 
     @Test
     void testSerialize() throws Exception {
-        var dto = new CommentDto(
+        var dto = new BookingInDto(
                 1L,
-                "Comment 1",
-                "User Name 1",
-                LocalDateTime.now()
+                LocalDateTime.now(),
+                LocalDateTime.now().plusDays(2L),
+                2L
         );
 
         var result = json.write(dto);
-
         assertThat(result).hasJsonPath("$.id");
-        assertThat(result).hasJsonPath("$.text");
-        assertThat(result).hasJsonPath("$.authorName");
-        assertThat(result).hasJsonPath("$.created");
+        assertThat(result).hasJsonPath("$.start");
+        assertThat(result).hasJsonPath("$.end");
+        assertThat(result).hasJsonPath("$.itemId");
 
         assertThat(result).extractingJsonPathNumberValue("$.id").isEqualTo(dto.getId().intValue());
-        assertThat(result).extractingJsonPathStringValue("$.text").isEqualTo(dto.getText());
-        assertThat(result).extractingJsonPathStringValue("$.authorName").isEqualTo(dto.getAuthorName());
-        assertThat(result).extractingJsonPathStringValue("$.created").startsWith(dto.getCreated().format(dtf));
+        assertThat(result).extractingJsonPathStringValue("$.start").startsWith(dto.getStart().format(dtf));
+        assertThat(result).extractingJsonPathStringValue("$.end").startsWith(dto.getEnd().format(dtf));
+        assertThat(result).extractingJsonPathNumberValue("$.itemId").isEqualTo(dto.getItemId().intValue());
     }
 }
