+++ common/src/test/java/shareit/comment/dto/CommentDtoTest.java
@@ -1,10 +1,11 @@
-package ru.practicum.shareit.booking.dto;
+package shareit.comment.dto;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.json.JsonTest;
 import org.springframework.boot.test.json.JacksonTester;
 import ru.practicum.shareit.Constants;
+import ru.practicum.shareit.comment.dto.CommentDto;
 
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
@@ -12,29 +13,30 @@ import java.time.format.DateTimeFormatter;
 import static org.assertj.core.api.AssertionsForInterfaceTypes.assertThat;
 
 @JsonTest
-class BookingInDtoTest {
+class CommentDtoTest {
     private final DateTimeFormatter dtf = Constants.DATE_TIME_FORMATTER;
     @Autowired
-    private JacksonTester<BookingInDto> json;
+    private JacksonTester<CommentDto> json;
 
     @Test
     void testSerialize() throws Exception {
-        var dto = new BookingInDto(
+        var dto = new CommentDto(
                 1L,
-                LocalDateTime.now(),
-                LocalDateTime.now().plusDays(2L),
-                2L
+                "Comment 1",
+                "User Name 1",
+                LocalDateTime.now()
         );
 
         var result = json.write(dto);
+
         assertThat(result).hasJsonPath("$.id");
-        assertThat(result).hasJsonPath("$.start");
-        assertThat(result).hasJsonPath("$.end");
-        assertThat(result).hasJsonPath("$.itemId");
+        assertThat(result).hasJsonPath("$.text");
+        assertThat(result).hasJsonPath("$.authorName");
+        assertThat(result).hasJsonPath("$.created");
 
         assertThat(result).extractingJsonPathNumberValue("$.id").isEqualTo(dto.getId().intValue());
-        assertThat(result).extractingJsonPathStringValue("$.start").startsWith(dto.getStart().format(dtf));
-        assertThat(result).extractingJsonPathStringValue("$.end").startsWith(dto.getEnd().format(dtf));
-        assertThat(result).extractingJsonPathNumberValue("$.itemId").isEqualTo(dto.getItemId().intValue());
+        assertThat(result).extractingJsonPathStringValue("$.text").isEqualTo(dto.getText());
+        assertThat(result).extractingJsonPathStringValue("$.authorName").isEqualTo(dto.getAuthorName());
+        assertThat(result).extractingJsonPathStringValue("$.created").startsWith(dto.getCreated().format(dtf));
     }
 }
