+++ common/src/test/java/shareit/item/dto/ItemDtoTest.java
@@ -1,31 +1,27 @@
-package ru.practicum.shareit.item.dto;
+package shareit.item.dto;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.json.JsonTest;
 import org.springframework.boot.test.json.JacksonTester;
-import ru.practicum.shareit.booking.dto.BookingItemDto;
-
-import java.util.ArrayList;
+import ru.practicum.shareit.item.dto.ItemDto;
 
 import static org.assertj.core.api.AssertionsForInterfaceTypes.assertThat;
 
 @JsonTest
-class ItemWithBookingDtoTest {
+class ItemDtoTest {
     @Autowired
-    private JacksonTester<ItemWithBookingDto> json;
+    private JacksonTester<ItemDto> json;
 
     @Test
     void testSerialize() throws Exception {
-        var dto = new ItemWithBookingDto(
+        var dto = new ItemDto(
                 1L,
                 "Item1",
                 "Item 1 description",
-                true,
-                new BookingItemDto(),
-                new BookingItemDto(),
-                new ArrayList<>(),
-                1L
+                false,
+                2L,
+                3L
         );
 
         var result = json.write(dto);
@@ -33,14 +29,13 @@ class ItemWithBookingDtoTest {
         assertThat(result).hasJsonPath("$.name");
         assertThat(result).hasJsonPath("$.description");
         assertThat(result).hasJsonPath("$.available");
-        assertThat(result).hasJsonPath("$.lastBooking");
-        assertThat(result).hasJsonPath("$.nextBooking");
-        assertThat(result).hasJsonPath("$.comments");
+        assertThat(result).hasJsonPath("$.requestId");
         assertThat(result).hasEmptyJsonPathValue("$.owner");
 
         assertThat(result).extractingJsonPathNumberValue("$.id").isEqualTo(dto.getId().intValue());
         assertThat(result).extractingJsonPathStringValue("$.name").isEqualTo(dto.getName());
         assertThat(result).extractingJsonPathStringValue("$.description").isEqualTo(dto.getDescription());
         assertThat(result).extractingJsonPathBooleanValue("$.available").isEqualTo(dto.getAvailable());
+        assertThat(result).extractingJsonPathNumberValue("$.requestId").isEqualTo(dto.getRequestId().intValue());
     }
 }
