+++ common/src/test/java/shareit/item/dto/ItemWithBookingDtoTest.java
@@ -1,26 +1,32 @@
-package ru.practicum.shareit.item.dto;
+package shareit.item.dto;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.json.JsonTest;
 import org.springframework.boot.test.json.JacksonTester;
+import ru.practicum.shareit.booking.dto.BookingItemDto;
+import ru.practicum.shareit.item.dto.ItemWithBookingDto;
+
+import java.util.ArrayList;
 
 import static org.assertj.core.api.AssertionsForInterfaceTypes.assertThat;
 
 @JsonTest
-class ItemDtoTest {
+class ItemWithBookingDtoTest {
     @Autowired
-    private JacksonTester<ItemDto> json;
+    private JacksonTester<ItemWithBookingDto> json;
 
     @Test
     void testSerialize() throws Exception {
-        var dto = new ItemDto(
+        var dto = new ItemWithBookingDto(
                 1L,
                 "Item1",
                 "Item 1 description",
-                false,
-                2L,
-                3L
+                true,
+                new BookingItemDto(),
+                new BookingItemDto(),
+                new ArrayList<>(),
+                1L
         );
 
         var result = json.write(dto);
@@ -28,13 +34,14 @@ class ItemDtoTest {
         assertThat(result).hasJsonPath("$.name");
         assertThat(result).hasJsonPath("$.description");
         assertThat(result).hasJsonPath("$.available");
-        assertThat(result).hasJsonPath("$.requestId");
+        assertThat(result).hasJsonPath("$.lastBooking");
+        assertThat(result).hasJsonPath("$.nextBooking");
+        assertThat(result).hasJsonPath("$.comments");
         assertThat(result).hasEmptyJsonPathValue("$.owner");
 
         assertThat(result).extractingJsonPathNumberValue("$.id").isEqualTo(dto.getId().intValue());
         assertThat(result).extractingJsonPathStringValue("$.name").isEqualTo(dto.getName());
         assertThat(result).extractingJsonPathStringValue("$.description").isEqualTo(dto.getDescription());
         assertThat(result).extractingJsonPathBooleanValue("$.available").isEqualTo(dto.getAvailable());
-        assertThat(result).extractingJsonPathNumberValue("$.requestId").isEqualTo(dto.getRequestId().intValue());
     }
 }
