--- gateway/src/main/java/ru/practicum/shareit/Constants.java
@@ -1,5 +1,0 @@
-package ru.practicum.shareit;
-
-public class Constants {
-    public static final String PAGE_SIZE_STRING = "20";
-}
