+++ gateway/src/main/java/ru/practicum/shareit/WebClientConfig.java
@@ -1,0 +1,23 @@
+package ru.practicum.shareit;
+
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.boot.web.client.RestTemplateBuilder;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.ComponentScan;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.http.client.HttpComponentsClientHttpRequestFactory;
+import org.springframework.web.client.RestTemplate;
+import org.springframework.web.util.DefaultUriBuilderFactory;
+
+@Configuration
+@ComponentScan
+public class WebClientConfig {
+    @Bean
+    public RestTemplate restTemplate(@Value("${shareit-server.url}") String serverUrl) {
+
+        return new RestTemplateBuilder()
+                .uriTemplateHandler(new DefaultUriBuilderFactory(serverUrl))
+                .requestFactory(HttpComponentsClientHttpRequestFactory::new)
+                .build();
+    }
+}
