+++ gateway/src/main/java/ru/practicum/shareit/booking/BookingClient.java
@@ -1,12 +1,8 @@
 package ru.practicum.shareit.booking;
 
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.boot.web.client.RestTemplateBuilder;
 import org.springframework.http.ResponseEntity;
-import org.springframework.http.client.HttpComponentsClientHttpRequestFactory;
 import org.springframework.stereotype.Service;
-import org.springframework.web.util.DefaultUriBuilderFactory;
+import org.springframework.web.client.RestTemplate;
 import ru.practicum.shareit.booking.dto.BookingInDto;
 import ru.practicum.shareit.booking.model.BookingRequestState;
 import ru.practicum.shareit.client.BaseClient;
@@ -17,29 +13,23 @@ import java.util.Map;
 public class BookingClient extends BaseClient {
     private static final String API_PREFIX = "/bookings";
 
-    @Autowired
-    public BookingClient(@Value("${shareit-server.url}") String serverUrl, RestTemplateBuilder builder) {
-        super(
-                builder
-                        .uriTemplateHandler(new DefaultUriBuilderFactory(serverUrl + API_PREFIX))
-                        .requestFactory(HttpComponentsClientHttpRequestFactory::new)
-                        .build()
-        );
+    public BookingClient(RestTemplate rest) {
+        super(rest);
     }
 
     public ResponseEntity<Object> createBooking(Long userId, BookingInDto bookingInDto) {
-        return post("", userId, bookingInDto);
+        return post(API_PREFIX, userId, bookingInDto);
     }
 
     public ResponseEntity<Object> updateBookingApproveStatus(Long userId, Long bookingId, String bookingStatus) {
         Map<String, Object> parameters = Map.of(
                 "approved", bookingStatus
         );
-        return patch("/" + bookingId + "?approved={approved}", userId, parameters, null);
+        return patch(API_PREFIX + "/" + bookingId + "?approved={approved}", userId, parameters, null);
     }
 
     public ResponseEntity<Object> findBookingById(Long userId, Long bookingId) {
-        return get("/" + bookingId, userId);
+        return get(API_PREFIX + "/" + bookingId, userId);
     }
 
     public ResponseEntity<Object> findAllBookingByUserAndState(Long userId, BookingRequestState state, Integer from, Integer size) {
@@ -48,7 +38,7 @@ public class BookingClient extends BaseClient {
                 "from", from,
                 "size", size
         );
-        return get("/?state={state}&from={from}&size={size}", userId, parameters);
+        return get(API_PREFIX + "/?state={state}&from={from}&size={size}", userId, parameters);
     }
 
     public ResponseEntity<Object> findAllBookingByOwnerAndState(Long ownerId, BookingRequestState state, Integer from, Integer size) {
@@ -57,6 +47,6 @@ public class BookingClient extends BaseClient {
                 "from", from,
                 "size", size
         );
-        return get("/owner?state={state}&from={from}&size={size}", ownerId, parameters);
+        return get(API_PREFIX + "/owner?state={state}&from={from}&size={size}", ownerId, parameters);
     }
 }
