+++ gateway/src/main/java/ru/practicum/shareit/item/ItemClient.java
@@ -1,12 +1,8 @@
 package ru.practicum.shareit.item;
 
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.boot.web.client.RestTemplateBuilder;
 import org.springframework.http.ResponseEntity;
-import org.springframework.http.client.HttpComponentsClientHttpRequestFactory;
 import org.springframework.stereotype.Service;
-import org.springframework.web.util.DefaultUriBuilderFactory;
+import org.springframework.web.client.RestTemplate;
 import ru.practicum.shareit.client.BaseClient;
 import ru.practicum.shareit.comment.dto.CommentInDto;
 import ru.practicum.shareit.item.dto.ItemDto;
@@ -17,14 +13,8 @@ import java.util.Map;
 public class ItemClient extends BaseClient {
     private static final String API_PREFIX = "/items";
 
-    @Autowired
-    public ItemClient(@Value("${shareit-server.url}") String serverUrl, RestTemplateBuilder builder) {
-        super(
-                builder
-                        .uriTemplateHandler(new DefaultUriBuilderFactory(serverUrl + API_PREFIX))
-                        .requestFactory(HttpComponentsClientHttpRequestFactory::new)
-                        .build()
-        );
+    public ItemClient(RestTemplate rest) {
+        super(rest);
     }
 
     public ResponseEntity<Object> findAllByUserId(Long userId, Integer from, Integer size) {
@@ -32,19 +22,19 @@ public class ItemClient extends BaseClient {
                 "from", from,
                 "size", size
         );
-        return get("/?from={from}&size={size}", userId, parameters);
+        return get(API_PREFIX + "/?from={from}&size={size}", userId, parameters);
     }
 
     public ResponseEntity<Object> createItem(Long userId, ItemDto itemDto) {
-        return post("", userId, itemDto);
+        return post(API_PREFIX + "", userId, itemDto);
     }
 
     public ResponseEntity<Object> findItemWithBookingById(Long userId, Long itemId) {
-        return get("/" + itemId, userId);
+        return get(API_PREFIX + "/" + itemId, userId);
     }
 
     public ResponseEntity<Object> updateItem(Long userId, Long itemId, ItemDto itemDto) {
-        return patch("/" + itemId, userId, itemDto);
+        return patch(API_PREFIX + "/" + itemId, userId, itemDto);
     }
 
     public ResponseEntity<Object> findItemsByQueryText(String queryText, Integer from, Integer size) {
@@ -53,10 +43,10 @@ public class ItemClient extends BaseClient {
                 "from", from,
                 "size", size
         );
-        return get("/search?text={text}&from={from}&size={size}", parameters);
+        return get(API_PREFIX + "/search?text={text}&from={from}&size={size}", parameters);
     }
 
     public ResponseEntity<Object> addCommentToItem(Long userId, Long itemId, CommentInDto commentInDto) {
-        return post("/" + itemId + "/comment", userId, commentInDto);
+        return post(API_PREFIX + "/" + itemId + "/comment", userId, commentInDto);
     }
 }
