+++ gateway/src/main/java/ru/practicum/shareit/request/RequestClient.java
@@ -1,12 +1,8 @@
 package ru.practicum.shareit.request;
 
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.boot.web.client.RestTemplateBuilder;
 import org.springframework.http.ResponseEntity;
-import org.springframework.http.client.HttpComponentsClientHttpRequestFactory;
 import org.springframework.stereotype.Service;
-import org.springframework.web.util.DefaultUriBuilderFactory;
+import org.springframework.web.client.RestTemplate;
 import ru.practicum.shareit.client.BaseClient;
 import ru.practicum.shareit.request.dto.RequestInDto;
 
@@ -16,18 +12,13 @@ import java.util.Map;
 public class RequestClient extends BaseClient {
     private static final String API_PREFIX = "/requests";
 
-    @Autowired
-    public RequestClient(@Value("${shareit-server.url}") String serverUrl, RestTemplateBuilder builder) {
-        super(
-                builder
-                        .uriTemplateHandler(new DefaultUriBuilderFactory(serverUrl + API_PREFIX))
-                        .requestFactory(HttpComponentsClientHttpRequestFactory::new)
-                        .build()
-        );
+    public RequestClient(RestTemplate rest) {
+        super(rest);
     }
 
+
     public ResponseEntity<Object> createItemRequest(Long userId, RequestInDto requestInDto) {
-        return post("", userId, requestInDto);
+        return post(API_PREFIX + "", userId, requestInDto);
     }
 
     public ResponseEntity<Object> findAllRequestByUserId(Long userId, Integer from, Integer size) {
@@ -35,11 +26,11 @@ public class RequestClient extends BaseClient {
                 "from", from,
                 "size", size
         );
-        return get("/?from={from}&size={size}", userId, parameters);
+        return get(API_PREFIX + "/?from={from}&size={size}", userId, parameters);
     }
 
     public ResponseEntity<Object> getRequestById(Long userId, Long requestId) {
-        return get("/" + requestId, userId);
+        return get(API_PREFIX + "/" + requestId, userId);
     }
 
     public ResponseEntity<Object> getPageableRequestById(Long userId, Integer from, Integer size) {
@@ -47,6 +38,6 @@ public class RequestClient extends BaseClient {
                 "from", from,
                 "size", size
         );
-        return get("/all?from={from}&size={size}", userId, parameters);
+        return get(API_PREFIX + "/all?from={from}&size={size}", userId, parameters);
     }
 }
