+++ gateway/src/main/java/ru/practicum/shareit/user/UserClient.java
@@ -1,12 +1,8 @@
 package ru.practicum.shareit.user;
 
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.boot.web.client.RestTemplateBuilder;
 import org.springframework.http.ResponseEntity;
-import org.springframework.http.client.HttpComponentsClientHttpRequestFactory;
 import org.springframework.stereotype.Service;
-import org.springframework.web.util.DefaultUriBuilderFactory;
+import org.springframework.web.client.RestTemplate;
 import ru.practicum.shareit.client.BaseClient;
 import ru.practicum.shareit.user.dto.UserDto;
 
@@ -16,38 +12,33 @@ import java.util.Map;
 public class UserClient extends BaseClient {
     private static final String API_PREFIX = "/users";
 
-    @Autowired
-    public UserClient(@Value("${shareit-server.url}") String serverUrl, RestTemplateBuilder builder) {
-        super(
-                builder
-                        .uriTemplateHandler(new DefaultUriBuilderFactory(serverUrl + API_PREFIX))
-                        .requestFactory(HttpComponentsClientHttpRequestFactory::new)
-                        .build()
-        );
+    public UserClient(RestTemplate rest) {
+        super(rest);
     }
+
 
     public ResponseEntity<Object> findAll(Integer from, Integer size) {
         Map<String, Object> parameters = Map.of(
                 "from", from,
                 "size", size
         );
-        return get("?from={from}&size={size}", null, parameters);
+        return get(API_PREFIX + "?from={from}&size={size}", null, parameters);
     }
 
 
     public ResponseEntity<Object> createUser(UserDto userDto) {
-        return post("", userDto);
+        return post(API_PREFIX, userDto);
     }
 
     public ResponseEntity<Object> findUserById(Long userId) {
-        return get("/" + userId);
+        return get(API_PREFIX + "/" + userId);
     }
 
     public ResponseEntity<Object> updateUser(Long userId, UserDto userDto) {
-        return patch("/" + userId, userDto);
+        return patch(API_PREFIX + "/" + userId, userDto);
     }
 
     public ResponseEntity<Object> deleteUserById(Long userId) {
-        return delete("/" + userId);
+        return delete(API_PREFIX + "/" + userId);
     }
 }
