--- server/src/main/java/ru/practicum/shareit/booking/model/BookingStatus.java
@@ -1,8 +1,0 @@
-package ru.practicum.shareit.booking.model;
-
-public enum BookingStatus {
-    WAITING,
-    APPROVED,
-    REJECTED,
-    CANCELED
-}
