+++ server/src/main/java/ru/practicum/shareit/booking/service/BookingService.java
@@ -1,6 +1,5 @@
 package ru.practicum.shareit.booking.service;
 
-import org.springframework.data.domain.Pageable;
 import ru.practicum.shareit.booking.dto.BookingInDto;
 import ru.practicum.shareit.booking.dto.BookingOutDto;
 import ru.practicum.shareit.booking.exception.BookingErrorException;
@@ -25,5 +24,5 @@ public interface BookingService {
 
     List<BookingOutDto> findAllBookingByOwnerIdAndItemId(Long owner, Long itemId) throws UserNotFoundException, ItemNotFoundException;
 
-    List<Booking> findAllBookingByUserIdAndItemId(Pageable pageable, Long userId, Long itemId, LocalDateTime dateTime);
+    List<Booking> findAllBookingByUserIdAndItemId(Long userId, Long itemId, LocalDateTime dateTime);
 }
