+++ server/src/main/resources/application.properties
@@ -6,11 +6,13 @@ spring.sql.init.mode=always
 #---
 spring.datasource.driver-class-name=org.postgresql.Driver
 #spring.datasource.url=jdbc:postgresql://locahost:5432/shareit
-#spring.datasource.username=shareit
-#spring.datasource.password=password
-spring.datasource.url=${SPRING_DATASOURCE_URL}
-spring.datasource.username=${POSTGRES_USER}
-spring.datasource.password=${POSTGRES_PASSWORD}
+spring.datasource.driverClassName=org.h2.Driver
+spring.datasource.url=jdbc:h2:mem:shareit
+spring.datasource.username=shareit
+spring.datasource.password=password
+#spring.datasource.url=${SPRING_DATASOURCE_URL}
+#spring.datasource.username=${POSTGRES_USER}
+#spring.datasource.password=${POSTGRES_PASSWORD}
 #---
 spring.config.activate.on-profile=ci,test
 spring.datasource.driverClassName=org.h2.Driver
